+++ tests/test_checker/test_noqa.py
@@ -152,6 +152,7 @@ SHOULD_BE_RAISED = types.MappingProxyType({
     'WPS439': 1,
     'WPS440': 1,
     'WPS441': 1,
+    'WPS442': 1,
 
     'WPS500': 1,
     'WPS501': 1,